+++ kamehouse-shell/bin/common/functions/kamehouse/kamehouse-functions.sh
@@ -106,13 +106,23 @@ loadDockerContainerEnv() {
 }
 loadDockerContainerEnv
 
-loadKamehouseShellPwd() {
-  source ${HOME}/.kamehouse/config/.shell/shell.pwd
+loadKamehouseSecrets() {
+  migrateShellPwdToKameHouseSecrets
+  source ${HOME}/.kamehouse/config/keys/.kamehouse-secrets.cfg
   if [ "$?" != "0" ]; then
-    log.error "Error importing ~/.kamehouse/config/.shell/shell.pwd"
+    log.error "Error importing ~/.kamehouse/config/keys/.kamehouse-secrets.cfg"
     exit 99
   fi
-  log.trace "Loaded ~/.kamehouse/config/.shell/shell.pwd"
+  log.trace "Loaded ~/.kamehouse/config/keys/.kamehouse-secrets.cfg"
+}
+
+migrateShellPwdToKameHouseSecrets() {
+  if [ ! -f "${HOME}/.kamehouse/config/keys/.kamehouse-secrets.cfg" && -f "${HOME}/.kamehouse/config/.shell/shell.pwd" ]; then
+    log.warn ".kamehouse-secrets.cfg doesn't exist. Migrating shell.pwd to .kamehouse-secrets.cfg"
+    mkdir -p ${HOME}/.kamehouse/config/keys
+    mv ${HOME}/.kamehouse/config/.shell/shell.pwd ${HOME}/.kamehouse/config/keys/.kamehouse-secrets.cfg
+    rm -r ${HOME}/.kamehouse/config/.shell
+  fi
 }
 
 parseHttpdPort() {