+++ kamehouse-shell/bin/common/sudoers/www-data/groot-get-config.sh
@@ -1,7 +1,7 @@
 #!/bin/bash
 
 main() {
-  cat ${HOME}/.kamehouse/config/.shell/shell.pwd | grep "MARIADB_PASS_KAMEHOUSE"
+  cat ${HOME}/.kamehouse/config/keys/.kamehouse-secrets.cfg | grep "MARIADB_PASS_KAMEHOUSE"
 }
 
 main "$@"