+++ kamehouse-shell/bin/kamehouse/build-kamehouse.sh
@@ -13,7 +13,7 @@ if [ "$?" != "0" ]; then
   exit 99
 fi
 
-loadKamehouseShellPwd
+loadKamehouseSecrets
 
 # Run the build on this script always from the current directory
 USE_CURRENT_DIR=true