+++ kamehouse-shell/bin/kamehouse/deploy-kamehouse-dev.sh
@@ -24,7 +24,7 @@ if [ "$?" != "0" ]; then
   echo -e "\033[1;36m$(date +%Y-%m-%d' '%H:%M:%S)\033[0;39m - [\033[1;31mERROR\033[0;39m] - \033[1;31mAn error occurred importing deployment-functions.sh\033[0;39m"
   exit 99
 fi
-loadKamehouseShellPwd
+loadKamehouseSecrets
 
 EXIT_CODE=${EXIT_SUCCESS}
 DEPLOYMENT_DIR=""