+++ kamehouse-shell/bin/kamehouse/exec-kamehouse-all-servers.sh
@@ -11,7 +11,7 @@ if [ "$?" != "0" ]; then
   echo -e "\033[1;36m$(date +%Y-%m-%d' '%H:%M:%S)\033[0;39m - [\033[1;31mERROR\033[0;39m] - \033[1;31mAn error occurred importing docker-functions.sh\033[0;39m"
   exit 99
 fi
-loadKamehouseShellPwd
+loadKamehouseSecrets
 
 SCRIPT=""
 SCRIPT_ARGS=""